CHANGES:small changes/fixes to config files FIXED: extractor some images still had transparency colours. FIXED: Error in code causing extractor to not work with images STARTED:Button Types need to continue with this feature

+++ src/Extractor/Settlers3/GFX/GFXLandscape.cpp
@@ -44,7 +44,7 @@ bool Extractor::Settlers3::GFXLandscape::SaveToFile(std::string location){
 		location += "/Landscape/";
 		Functions::CreateDir(location);
 		for(unsigned short i = 0; i < this->count; i++){
-			if(this->images[i] != NULL) this->images[i]->SaveToFile(location + Functions::ToString(i));
+			if(this->images[i] != NULL) this->images[i]->SaveToFile(location + Functions::ToString((int)i));
 		}
 		return true;
 	}
